Fix to_geoframe and delegate union_all and crosses for more examples
